deps/v8/test/inspector/runtime/run-if-waiting-for-debugger-expected.txt


Running test: testTwoSessions
Tests Runtime.runIfWaitingForDebugger
session 1 resumed
session 2 resumed
execution resumed

Running test: testSessionDisconnect
Tests Runtime.runIfWaitingForDebugger
session 1 resumed
session 2 disconnected
execution resumed
